requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
python-docx==0.8.11
opencv-python==4.8.1.78

# Vector database and embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# LLM integration
google-generativeai==0.3.2

# Data processing
pandas==2.1.3
numpy==1.24.3

# Web interface
streamlit==1.28.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1